requirements.txt
pip
autopep8
Flask==3.0.3
gunicorn==22.0.0
Werkzeug==3.0.6
python-telegram-bot[job-queue]
requests
python-dotenv